[feat] 회원 기본 CRUD 생성, member ID Long 타입 변경으로 인한 수정

+++ src/main/java/store/teabliss/member/controller/MemberController.java
@@ -24,9 +24,9 @@ public class MemberController {
 
     @PostMapping("/sign-up")
     @Operation(summary = "회원 가입", description = "회원 등록 API")
-    public ResponseEntity<Integer> signUp(@RequestBody MemberSignUpDto memberSignUpDto) {
+    public ResponseEntity<Long> signUp(@RequestBody MemberSignUpDto memberSignUpDto) {
 
-        int memId = memberService.createMember(memberSignUpDto);
+        long memId = memberService.createMember(memberSignUpDto);
 
         return ResponseEntity.ok(memId);
     }
